+++ lib/remote_factory_girl.rb
@@ -1,5 +1,6 @@
 require "remote_factory_girl/version"
 require 'remote_factory_girl/config'
+require 'remote_factory_girl/remotes_config'
 require 'remote_factory_girl/http'
 require 'remote_factory_girl/config_applier'
 require 'remote_factory_girl/hash_to_dot'
@@ -23,32 +24,35 @@ module RemoteFactoryGirl
     end
   end
 
-  def self.configure(config = Config, &block)
+  def self.configure(remote_name = remotes_config.default_remote_name, config = Config, &block)
     configuration = Config.new
     yield(configuration)
-    self.config = configuration
+    remotes_config.remotes[remote_name] = configuration
   end
 
   def self.factories(params = {}, http = Http)
-    factory = RemoteFactoryGirl.new(config)
+    config_for_remote = config(remotes_config.current_remote)
+    factory = RemoteFactoryGirl.new(config_for_remote)
     factory.factories(params, http).to_hash
   end
 
   def self.create(factory_name, attributes = {}, config_applier = ConfigApplier, http = Http)
-    factory = RemoteFactoryGirl.new(config)
+    config_for_remote = config(remotes_config.current_remote)
+    factory = RemoteFactoryGirl.new(config_for_remote)
     factory.create(factory_name, attributes, http)
-    config_applier.apply_config(factory.response.to_hash, config.to_hash)
+    config_applier.apply_config(factory.response.to_hash, config_for_remote.to_hash)
   end
 
-  def self.config
-    @config
+  def self.with_remote(remote_name = remotes_config.default_remote_name)
+    remotes_config.current_remote = remote_name
+    self
   end
 
-  def self.config=(config)
-    @config = config
+  def self.config(remote_name = remotes_config.default_remote_name)
+    remotes_config.remotes[remote_name]
   end
 
-  def self.reset(config = Config.new)
-    self.config = config
+  def self.remotes_config
+    @remotes_config ||= RemotesConfig.new
   end
 end