+++ spec/integration/remote_factory_girl_spec.rb
@@ -1,8 +1,9 @@
 require 'remote_factory_girl'
+require 'spec_helper'
 
 describe RemoteFactoryGirl do
 
-  before { RemoteFactoryGirl.reset }
+  before { RemoteFactoryGirl.remotes_config.reset }
 
   describe 'configuration' do
     it 'should be configured with correct defaults' do
@@ -49,6 +50,34 @@ describe RemoteFactoryGirl do
       RemoteFactoryGirl.config.https = true
       expect(RemoteFactoryGirl.config.https).to eq true
     end
+
+    context 'when configuring multiple remotes' do
+
+      before do
+        configure_remote_factory_girl(remote_name: :travis,
+                                      host: 'localhost',
+                                      port: 3000,
+                                      end_point: '/remote_factory_girl/travis/home')
+        configure_remote_factory_girl(remote_name: :casey,
+                                      host: 'over_the_rainbow',
+                                      port: 6000,
+                                      end_point: '/remote_factory_girl/casey/home')
+      end
+
+      it 'should return configuration for remote "travis"' do
+        expect(RemoteFactoryGirl.config(:travis).home).to eq({:host      => 'localhost',
+                                                              :port      => 3000,
+                                                              :end_point => '/remote_factory_girl/travis/home'})
+        expect(RemoteFactoryGirl.config(:travis).return_with_root).to eq(true)
+        expect(RemoteFactoryGirl.config(:travis).return_response_as).to eq(:as_hash)
+      end
+
+      it 'should return configuration for remote "casey"' do
+        expect(RemoteFactoryGirl.config(:casey).home).to eq({:host      => 'over_the_rainbow',
+                                                             :port      => 6000,
+                                                             :end_point => '/remote_factory_girl/casey/home'})
+      end
+    end
   end
 
   describe 'errors' do
@@ -71,15 +100,114 @@ describe RemoteFactoryGirl do
     end
 
     describe '.factories' do
+      context 'when multiple remotes are configured' do
 
-      before { RemoteFactoryGirl.config.home[:host] = 'localhost' }
+        before do
+          configure_remote_factory_girl(remote_name: :travis,
+                                        host: 'localhost',
+                                        port: 3000,
+                                        end_point: '/remote_factory_girl/travis/home')
+          configure_remote_factory_girl(remote_name: :casey,
+                                        host: 'over_the_rainbow',
+                                        port: 6000,
+                                        end_point: '/remote_factory_girl/casey/home')
+        end
 
-      it 'should return all factories available' do
-        expect(RemoteFactoryGirl.factories).to match_array(['user', 'user_admin'])
+        context 'for remote "travis"' do
+          it 'should return all factories available' do
+            expect(RemoteFactoryGirl.with_remote(:travis).factories).to match_array(['user', 'user_admin'])
+          end
+
+          it 'should be configured to send HTTP request to remote "travis"' do
+            remote_factory_girl = RemoteFactoryGirl.with_remote(:travis)
+
+            expect(remote_factory_girl).to receive(:factories) do
+              expect(
+                remote_factory_girl.remotes_config.current_remote
+              ).to eq(:travis)
+              expect(
+                remote_factory_girl.config(:travis).home_url
+              ).to eq('http://localhost:3000/remote_factory_girl/travis/home')
+            end
+            remote_factory_girl.factories
+          end
+        end
+
+        context 'for remote "casey"' do
+          it 'should return all factories available' do
+            expect(RemoteFactoryGirl.with_remote(:casey).factories).to match_array(['user', 'user_admin'])
+          end
+
+          it 'should be configured to send HTTP request to remote "casey"' do
+            remote_factory_girl = RemoteFactoryGirl.with_remote(:casey)
+
+            expect(remote_factory_girl).to receive(:factories) do
+              expect(
+                remote_factory_girl.remotes_config.current_remote
+              ).to eq(:casey)
+              expect(
+                remote_factory_girl.config(:casey).home_url
+              ).to eq('http://over_the_rainbow:6000/remote_factory_girl/casey/home')
+            end
+            remote_factory_girl.factories
+          end
+        end
+      end
+
+      context 'when configured with remote "default"' do
+
+        before do
+          configure_remote_factory_girl(host: 'not_configured_with_name',
+                                        port: 9000)
+        end
+
+        it 'should return all factories available' do
+          expect(RemoteFactoryGirl.factories).to match_array(['user', 'user_admin'])
+        end
+
+        it 'should be configured to send HTTP request to remote "default"' do
+          remote_factory_girl = RemoteFactoryGirl
+
+          expect(remote_factory_girl).to receive(:factories) do
+            expect(
+              remote_factory_girl.remotes_config.current_remote
+            ).to eq(:default)
+            expect(
+              remote_factory_girl.config.home_url
+            ).to eq('http://not_configured_with_name:9000/remote_factory_girl/home')
+          end
+          remote_factory_girl.factories
+        end
       end
     end
 
     describe '.create' do
+      context 'when configured with multiple remotes' do
+        before do
+          configure_remote_factory_girl(remote_name: :travis,
+                                        host: 'localhost',
+                                        port: 3000,
+                                        end_point: '/remote_factory_girl/travis/home',
+                                        return_with_root: false)
+          configure_remote_factory_girl(remote_name: :casey,
+                                        host: 'over_the_rainbow',
+                                        port: 6000,
+                                        end_point: '/remote_factory_girl/casey/home',
+                                        return_response_as: :dot_notation,
+                                        return_with_root: false)
+
+        end
+
+        it 'should be able to create a factory with "travis" remote' do
+          user = RemoteFactoryGirl.with_remote(:travis).create(:user)
+          expect(user['first_name']).to eq('Sam')
+        end
+
+        it 'should be able to create a factory with "casey" remote' do
+          user = RemoteFactoryGirl.with_remote(:casey).create(:user)
+          expect(user.first_name).to eq('Sam')
+        end
+      end
 
       describe 'default .home' do
 
@@ -109,11 +237,10 @@ describe RemoteFactoryGirl do
       end
 
       it 'should not return root hash key and should return an object that responds to dot notation' do
-        RemoteFactoryGirl.configure do |config|
-          config.home               = { :host => 'localhost' }
-          config.return_response_as = :dot_notation
-          config.return_with_root   = false
-        end
+        configure_remote_factory_girl(host: 'localhost',
+                                      port: 3000,
+                                      return_response_as: :dot_notation,
+                                      return_with_root: false)
         user = RemoteFactoryGirl.create(:user)
         expect(user.first_name).to eq('Sam')
       end