+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### 1.1.0
+1. Configure and create test data in multiple "home"/"remote" applications
+2. Remove RemoteFactoryGirl.reset. Instead use, RemoteFactoryGirl.remotes_config.reset
 ### 1.0.2 - Aug. 19, 2014
 1. Update specs to use Rspec '3.0.0' syntax